Fix launch template optional API params
When using a launch template many of the params in the EC2 launch API
call are optional, but fail if you given an empty value, they need to
not be presemt. e.g. ImageID='' will fail. Therefore we only add them to
params if set.

The higher level validation on config handles start vs stop required
inputs.